// client/src/ui/MovieSearch/types.tsx
export type MovieData = {
  actors: string
  country: string
  director: string
  genre: string
  metascore: string
  plot: string
  poster: string
  rated: string
  // ratings: { Source: string; Value: string }[]
  released: string
  runtime: string
  title: string
  type: string // Type: movie
  writer: string
  year: string
  imdbID: string
  imdbRating: string
}